Fix revision number
